feat: refactor structure
