+++ Project/File/bill.txt
@@ -1,0 +1,3 @@
+customerId1,Nguyen Van A,123 Street, District 1, Ho Chi Minh City,nguyenvana@example.com,Laptop A,1,Maker A,1000.0,Laptop,2.5,Model A,24,2023-01-01,Intel i7,16,512,NVIDIA GTX 1650,Black,5,15.6,Windows 10,4,65.0,salerId1,salerName1,salerEmail1,salerPhoneNumber1,salerPosition1,5000.0,0.05,0987654321,billId1,2023-01-01,1500.0
+customerId2,Le Thi B,456 Avenue, District 2, Ho Chi Minh City,lethib@example.com,Laptop B,2,Maker B,1200.0,Laptop,2.0,Model B,12,2023-01-02,Intel i5,8,256,NVIDIA GTX 1650,Silver,4,14.0,Windows 10,4,65.0,salerId2,salerName2,salerEmail2,salerPhoneNumber2,salerPosition2,6000.0,0.07,0912345678,billId2,2023-02-01,1800.0
+customerId3,Tran Van C,789 Boulevard, District 3, Ho Chi Minh City,tranvanc@example.com,Laptop A,1,Maker A,1000.0,Laptop,2.5,Model A,24,2023-01-01,Intel i7,16,512,NVIDIA GTX 1650,Black,5,15.6,Windows 10,4,65.0,salerId3,salerName3,salerEmail3,salerPhoneNumber3,salerPosition3,7000.0,0.08,0934567890,billId3,2023-03-01,2000.0
