+++ Project/File/staff.txt
@@ -1,0 +1,2 @@
+E001,John Doe,johndoe@example.com,0123456789,Manager,10000.0,IT,1.5
+E002,Jane Smith,janesmith@example.com,0987654321,Saler,3000.0,6,2
